Refactor out the stdweb functionality into different modules
